+++ cmd/wyze-bridge/main.go
@@ -28,7 +28,7 @@ import (
 )
 
 // Version is set at build time via ldflags.
-var Version = "4.6.1"
+var Version = "4.6.2"
 
 func main() {
 	// Load configuration